skip existing, handle errors, log, correct file list. single-threaded downloader done, but it is slow.

+++ abide_brain_age/tools/download_abide_freesurfer_data.py
@@ -5,36 +5,47 @@
 # Written by Tim Schäfer, 2019-06-05
 
 
-import os
+import os, sys
 import pandas as pd
 import urllib.request
 import urllib.error
 import shutil
 import errno
 # Download the file from `url` and save it locally under `file_name`:
-
+from os.path import expanduser
 
 def get_abide():
-    local_base_dir = "./abide_data/"
+    #local_base_dir = "./abide_data/"
+    local_base_dir = os.path.join(expanduser("~"), "data", "abide", "structural")
     #required_files_relative_to_subject_dir = ["surf/lh.white", "surf/rh.white", "surf/notthere"]
     required_files_relative_to_subject_dir = get_fs_subject_filenames()
+
+    for file_idx, file in enumerate(required_files_relative_to_subject_dir):
+        print("%d: %s" % (file_idx+1, file))
+
+    diagnose_output("............................................ssssss......ssssss......ssssss......ssssss......ssssss......ssssss", required_files_relative_to_subject_dir)
 
     print("Downloading ABIDE I structural data to local directory '%s'." % (local_base_dir))
     print("Will download %d files per subject." % (len(required_files_relative_to_subject_dir)))
 
-    download_abide_structural_files_to(local_base_dir, required_files_relative_to_subject_dir)
+    download_abide_structural_files_to(local_base_dir, required_files_relative_to_subject_dir, skip_existing=True)
 
 
 def get_fs_subject_filenames():
     """Generates the file paths of all important FreeSurfer files for a subject, relative to its subject dir."""
     files = []
     files = files + _get_both_hemi_files_in_dir("surf", ["white", "pial", "inflated", "orig", "smoothwm", "sphere", "sphere.reg", "jacobian_white", "thickness", "area", "curv", "volume", "sulc"])
-    files = files + _get_both_hemi_files_in_dir("stats", ["aparc.stats", "aparc.a2009s.stats", "aparc.DKTatlas.stats", "aparc.pial.stats", "curv.stats"])
+    files = files + _get_both_hemi_files_in_dir("stats", ["aparc.stats", "aparc.a2009s.stats"])
     files = files + _get_files_in_subdir("stats", ["aseg.stats"])
-    files = files + _get_files_in_subdir("mri", ["brain.mgz", "brainmask.mgz", "orig.mgz", "T1.mgz", "asge.mgz", "aparc.a2009s+aseg.mgz", "aparc+aseg.mgz", "aparc.DKTatlas+aseg.mgz"])
+    files = files + _get_files_in_subdir("mri", ["brain.mgz", "brainmask.mgz", "orig.mgz", "T1.mgz", "aseg.mgz"])
     files = files + _get_files_in_subdir("mri/transforms", ["talairach.m3z", "talairach.xfm"])
     files = files + _get_surf_both_hemi_fsaverge_mappings(["area", "area.pial", "sulc", "thickness", "curv", "volume"])
     return files
+
+def diagnose_output(out_str, files):
+    for idx, c in enumerate(out_str):
+        if c != ".":
+            print("%s: %s" % (c, files[idx]))
 
 def _get_both_hemi_files_in_dir(subdir, file_list):
     files_lh = [("%s/lh.%s" % (subdir, f)) for f in file_list]
@@ -58,8 +69,7 @@ def _get_surf_both_hemi_fsaverge_mappings(file_name_parts_list, subdir="surf"):
     return files
 
 
-
-def download_abide_structural_files_to(local_base_dir, required_files_relative_to_subject_dir, verbose=True):
+def download_abide_structural_files_to(local_base_dir, required_files_relative_to_subject_dir, verbose=True, skip_existing=False):
     if not os.path.exists(local_base_dir):
         raise IOError("Local base directory '%s' does not exist, please create it first." % (local_base_dir))
     md = load_abide_metadata()
@@ -73,17 +83,17 @@ def download_abide_structural_files_to(local_base_dir, required_files_relative_t
         print("Received meta data on %d subjects. Will download %d * %d = %d files in total." % (num_ids, num_ids, len(required_files_relative_to_subject_dir), num_files_total))
 
     base_url = "https://s3.amazonaws.com/fcp-indi/data/Projects/ABIDE_Initiative/Outputs/freesurfer/5.1/"
-
     ok_files = []
     error_files = []
     error_full_urls = []
     error_messages = []
+    skipped_files = []
 
     for idx, fid in file_ids.items():
         if fid == "no_filename":
             continue
-        if idx % 3 == 0:
-            print("At subject number %d of %d: encountered problems with %d of %d files so far." % (idx, num_ids, len(error_files), len(error_files) + len(ok_files)))
+        if idx % 10 == 0:
+            print("At subject number %d of %d: encountered problems with %d of %d files so far, skipped %d." % (idx, num_ids, len(error_files), len(error_files) + len(ok_files), len(skipped_files)))
         if verbose:
             print("[%s]" % (fid), end="", flush=True)
         for fpath in required_files_relative_to_subject_dir:
@@ -93,24 +103,38 @@ def download_abide_structural_files_to(local_base_dir, required_files_relative_t
             file_name = fparts[-1]
             if len(fparts) > 1:
                 extra_dirs = fparts[0:-1]
-            code, full_url, msg = download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir)
-            if code != 0:
+            code, full_url, msg = download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir, skip_existing=skip_existing)
+            if code == 1:
                 error_files.append(fpath)
                 error_full_urls.append(full_url)
                 error_messages.append(msg)
-            else:
+                if verbose:
+                    print("E", end="", flush=True)
+            elif code == -1:
+                skipped_files.append(fpath)
+                if verbose:
+                    print("s", end="", flush=True)
+            elif code == 0:
                 ok_files.append(fpath)
-            if verbose:
-                print(".", end="", flush=True)
+                if verbose:
+                    print(".", end="", flush=True)
+            else:
+                print("Unhandled return code received from download function, exiting.")
+                sys.exit(1)
+
         if verbose:
             print("")
     if error_files:
         print("Encountered problems with %d of the %d files:" % (len(error_files), num_files_total))
         for idx, fname in enumerate(error_files):
-            print("file %s : url '%s' : error '%s'" % (error_files[idx], error_full_urls[idx], error_messages[idx]))
+            print("-error file %s : url '%s' : error '%s'" % (error_files[idx], error_full_urls[idx], error_messages[idx]))
+    if skipped_files:
+        print("Skipped %d of the %d files:" % (len(skipped_files), num_files_total))
+        for idx, fname in enumerate(skipped_files):
+            print("-skipped file %s" % (skipped_files[idx]))
 
 
-def download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir):
+def download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir, skip_existing=False):
     """
     Download a remote file from a base url to a local folder, and create the equivalent directory structure in a local base dir.
 
@@ -123,6 +147,10 @@ def download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir):
     local_dir = os.path.join(local_base_dir, os.path.join(*extra_dirs))
     mkdir_p(local_dir)
     local_file = os.path.join(local_dir, file_name)
+
+    if skip_existing and os.path.isfile(local_file):
+        return -1, full_url, "skipped on user request: local file exists"
+
     try:
         with urllib.request.urlopen(full_url) as response, open(local_file, 'wb') as out_file:
             shutil.copyfileobj(response, out_file)
@@ -130,6 +158,11 @@ def download_file_to_local_dir(base_url, extra_dirs, file_name, local_base_dir):
         return 1, full_url, str(h_err.code)
     except urllib.error.URLError as u_err:
         return 1, full_url, str(u_err.reason)
+    except Exception as e:
+        if hasattr(e, 'message'):
+            return 1, full_url, e.message
+        return 1, full_url, str(e)
+
     return 0, full_url, ""
